feat: improved handling of rpc creds for Join Market.
We now use dedicated rpc creds for JM which can easily be refreshed by clicking "Configure JM". This will trigger a cookie file to be written which consists of rpcuser:rpcpassword, it is stored in ~/.fullynoded/JoinMarket/.temp/.cookie, if you edit the creds you will need to generate an rpcauth string with your creds and add it to the bitcoin.conf too.